created room when accepting request and get member by userid

+++ models/chatmember.go
@@ -2,7 +2,7 @@ package models
 
 type ChatMember struct {
 	Base
-	UserID int   `json:"user_id"`
+	UserID int64 `json:"user_id"`
 	RoomID int64 `json:"room_id"`
 }
